feat(js-algo-ds/dynamic-programming): add dynamic programming docs, implementations and add helpers
