+++ JS Algorithms & Data Structures/helpers/performance-timer.js
@@ -1,0 +1,9 @@
+const performanceTimer = (func) => {
+  let start = performance.now();
+  func;
+  let end = performance.now();
+
+  console.log(`Execution time: ${end - start} ms`);
+};
+
+export { performanceTimer };
